create  separate folder for slice
