+++ src/redux/Store.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { configureStore } from '@reduxjs/toolkit'
-import counterReducer from './CounterSlice'
+import counterReducer from '../slices/CounterSlice'
 
 
 export const Store = configureStore({
